Change cache hit/read counter to use long. spunkiie's server would probably max out the integer cap :) (2.147 billion)

+++ src/main/java/com/griefcraft/util/Performance.java
@@ -66,7 +66,7 @@ public class Performance {
     /**
      * @return the average of a value
      */
-    public static double getAverage(int value) {
+    public static double getAverage(long value) {
         return (double) value / getTimeRunningSeconds();
     }
